requirements.txt
django==5.0.4
aiogram==3.4.1
python-decouple==3.8
channels==4.1.0
psycopg2-binary==2.9.9
daphne==4.1.0
whitenoise==6.6.0
requests==2.31.0